Move manifest controller code to pkg/controller/manifest
