+++ controllers/addon_controller.go
@@ -23,9 +23,9 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 
 	boundlessv1alpha1 "github.com/mirantiscontainers/boundless-operator/api/v1alpha1"
+	"github.com/mirantiscontainers/boundless-operator/pkg/controllers/manifest"
 	"github.com/mirantiscontainers/boundless-operator/pkg/event"
 	"github.com/mirantiscontainers/boundless-operator/pkg/helm"
-	"github.com/mirantiscontainers/boundless-operator/pkg/manifest"
 )
 
 const (
